--- Practice/Topic1/Exercise1_2_3_4/src/Cash.java
@@ -1,4 +1,0 @@
-
-public class Cash extends PaymentType {
-//Cash: no information is required. but it is a PaymentType.
-}
