--- Practice/Topic1/Exercise1_2_3_4/src/CreditCard.java
@@ -1,14 +1,0 @@
-
-public class CreditCard extends PaymentType{
-
-	// private instance variables
-	private String name;
-	private int number;
-
-	// Constructor
-	public CreditCard(String name, int number) {
-		this.name = name;
-		this.number = number;
-	}
-	
-}
