--- Practice/Topic1/Exercise1_2_3_4/src/PaymentType.java
@@ -1,5 +1,0 @@
-
-/**Abstract Class*/
-public abstract class PaymentType {
-
-}
