--- Practice/Topic1/Exercise1_2_3_4/src/Paypal.java
@@ -1,15 +1,0 @@
-
-public class Paypal extends PaymentType{
-
-	// private instance variables
-	private String email;
-	private String password;
-
-	// Constructor
-	public Paypal(String email, String password) {
-		super();
-		this.email = email;
-		this.password = password;
-	}
-		
-}
