+++ Practice/Topic1/Exercise1_2_3_4/src/com/mfidececchi/topic2/Cash.java
@@ -1,0 +1,11 @@
+package com.mfidececchi.topic2;
+
+public class Cash extends PaymentType {
+	// Cash: no information is required. but it is a PaymentType.
+
+	@Override
+	public String toString() {		
+		return "Cash @";
+	}
+
+}
