+++ Practice/Topic1/Exercise1_2_3_4/src/com/mfidececchi/topic2/CreditCard.java
@@ -1,0 +1,29 @@
+package com.mfidececchi.topic2;
+
+public class CreditCard extends PaymentType {
+
+	// private instance variables
+	private String name;
+	private int number;
+
+	// Constructor
+	public CreditCard(String name, int number) {
+		this.name = name;
+		this.number = number;
+	}
+
+	// getters
+	public String getName() {
+		return name;
+	}
+
+	public int getNumber() {
+		return number;
+	}
+
+	@Override
+	public String toString() {
+		return "CreditCard @ "+getName()+" "+getNumber();		
+	}
+
+}
