+++ Practice/Topic1/Exercise1_2_3_4/src/com/mfidececchi/topic2/PaymentLine.java
@@ -1,3 +1,4 @@
+package com.mfidececchi.topic2;
 
 public class PaymentLine {
