+++ Practice/Topic1/Exercise1_2_3_4/src/com/mfidececchi/topic2/PaymentType.java
@@ -1,0 +1,10 @@
+package com.mfidececchi.topic2;
+
+/** Abstract Class */
+public abstract class PaymentType {
+
+	// method to be inherited and overwritten in the subclasses
+	// the purpose is to return the payment type description
+	public abstract String toString();
+
+}
