+++ Practice/Topic1/Exercise1_2_3_4/src/com/mfidececchi/topic2/TotalCalculator.java
@@ -1,3 +1,4 @@
+package com.mfidececchi.topic2;
 import java.util.ArrayList;
 
 import com.sun.org.apache.xerces.internal.impl.xs.SubstitutionGroupHandler;
