+++ Practice/Topic1/Exercise1_2_3_4/src/com/mfidececchi/topic2/Transaction.java
@@ -1,3 +1,4 @@
+package com.mfidececchi.topic2;
 import java.time.Instant;
 import java.util.Date;
